test(harness): move live Gemini chat tests to chat_live_gemini suite
Per audit 5/8c: TestGoogleAIStudioGemini, the live image-gen trio
(sync/async/async-stream), thinking, finish_reason, url_context, grounding,
empty function-call args, tool use (+ its get_current_weather helper),
system-without-user message, openai-web-search-to-google-search. The
tool-call prompt-factory golden formerly hiding as the class override stays
behind as test_gemini_tool_call_invoke_no_arguments (fixture inlined, with a
real assertion instead of a bare print). Context-caching transforms,
image-config/size-mapping goldens, thinking param-mapping, exception mapping,
unicode args, and test_claude_tool_use_with_gemini stay as unit tests;
test_gemini_embedding stays (non-chat, Sameer)
